\name{datashield.heatmap.plot.combine}
\alias{datashield.heatmap.plot.combine}
\title{Generates a heatmap plot for merged datasets}
\usage{
  datashield.heatmap.plot.combine(opals, xvect, yvect)
}
\arguments{
  \item{opals}{character strings that represent the URL of
  the servers where the study datasets are stored.}

  \item{xvect}{a numerical vector}

  \item{yvect}{a numerical vector}
}
\value{
  a heatmap plot
}
\description{
  Generates a heatmap plot for merged datasets
}

